+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the WooCommerce Local Pickup Costs plugin will be documen
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] - 2025-01-XX
+
+### Fixed
+- Fixed settings save to properly read array-formatted POST data
+- Corrected how location costs are retrieved from `$_POST['wc_lpc_location_costs']` array
+- Settings now properly save and persist
+
 ## [1.0.5] - 2025-01-XX
 
 ### Fixed